FE 70% done
