+++ app/config/routes.js
@@ -2,12 +2,12 @@ import React from 'react';
 import ReactRouter, { Route, IndexRoute } from 'react-router';
 import App from '../containers/App';
 import DashboardCon from '../containers/DashboardCon';
-import Schedules from '../containers/Schedules';
+import SchedulesCon from '../containers/SchedulesCon';
 import CalendarCon from '../containers/CalendarCon';
 import AnnounceCon from '../containers/AnnounceCon';
 import ChatBoxCon from '../containers/ChatBoxCon';
 import GuestCon from '../containers/GuestCon';
-import Register from '../containers/Register';
+import RegisterCon from '../containers/RegisterCon';
 import LoginCon from '../containers/LoginCon';
 
 const routes = (
@@ -15,10 +15,10 @@ const routes = (
             <IndexRoute component={GuestCon}/>
             <Route path="dashboard" component={DashboardCon}/>
             <Route path="announcements" component={AnnounceCon}/>
-            <Route path="schedules" component={Schedules}/>
+            <Route path="schedules" component={SchedulesCon}/>
             <Route path="event" component={CalendarCon}/>
             <Route path="chat" component={ChatBoxCon}/>
-            <Route path="register" component={Register} />
+            <Route path="register" component={RegisterCon} />
             <Route path="login" component={LoginCon} />
         </Route>
 );
