+++ app/containers/SchedulesCon.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import FormSched from '../components/forms/FormSched';
+import Schedules from '../components/schedules/Schedules';
+import Data from '../data';
+
+const SchedulesCon = () => {
+
+  return (
+    <div>
+      <FormSched />
+      <Schedules />
+    </div>
+  );
+};
+
+export default SchedulesCon;
